minor changes in internal queue stats collection

+++ src/server/core/dbwrite.cpp
@@ -216,7 +216,7 @@ static THREAD_RESULT THREAD_CALL IDataWriteThread(void *arg)
 				if (hStmt != NULL)
 				{
 					DBBind(hStmt, 1, DB_SQLTYPE_INTEGER, rq->dciId);
-					DBBind(hStmt, 2, DB_SQLTYPE_INTEGER, rq->timestamp);
+					DBBind(hStmt, 2, DB_SQLTYPE_INTEGER, (INT64)rq->timestamp);
 					DBBind(hStmt, 3, DB_SQLTYPE_INTEGER, rq->value, DB_BIND_STATIC);
 					success = DBExecute(hStmt);
 					DBFreeStatement(hStmt);
